better name for CSV with PacBio CCS mutations
Previoiusly this output file was called `_aligned_with_ref_annotation.csv`.

It seems like the main purpose of this file is actually to store the mutations called per CCS for downstream processing. Accordingly, I have re-named it to `_mutations_by_ccs.csv`, which better reflects this purpose of the file.